🐛 Fix datepicker saving wrong date
In some situations selected dates were being saved
as the day before was intended. This was due to
the datepicker returning a unix epoch representing
the current date at midnight, which was then being
parsed with `new Date(date).toISOString()` which
would add the current browser's timezone and then
adjust to UTC which could move it to the day
before.

Now we use dayjs' format method which simply
outputs the date in the format we want without
adjusting for timezones.